Splits runtime built in functions into semantic structure: arithmetic, financial, logical, etc.
